requirements.txt
Eve==1.1.2
pymongo==3.11.0
PyJWT==1.7.1

Flask~=3.0.3
passlib~=1.7.4
Flask-Cors~=5.0.0